fix(solver): split conditional branch relation helper
